1.0.5

+++ src/main/java/br/com/projetosalura/AppTabelaFipe/demo/models/IModelType.java
@@ -2,5 +2,5 @@ package br.com.projetosalura.AppTabelaFipe.demo.models;
 
 public interface IModelType {
     String getName();
-    int getCode();
+    long getCode();
 }
